[WIN32SS] Scrollbars - Integer underflow/wraparound

+++ win32ss/user/ntuser/scrollbar.c
@@ -597,7 +597,8 @@ co_IntSetScrollInfo(PWND Window, INT nBar, LPCSCROLLINFO lpsi, BOOL bRedraw)
    if (lpsi->fMask & (SIF_RANGE | SIF_PAGE | SIF_DISABLENOSCROLL))
    {
       new_flags = Window->pSBInfo->WSBflags;
-      if (Info->nMin >= (int)(Info->nMax - max(Info->nPage - 1, 0)))
+      /* Cast Info->nPage to int to prevent wraparound/underflow. Fixes CORE-16491 */
+      if (Info->nMin >= (int)(Info->nMax - max((int)Info->nPage - 1, 0)))
       {
          /* Hide or disable scroll-bar */
          if (lpsi->fMask & SIF_DISABLENOSCROLL)
